requirements.txt
Flask
Flask-Cors
Pillow
reportlab
pypdf
gunicorn
cryptography>=3.1